etc: CourseResponseDtoDataTypes 공통 타입으로 이동
